+++ NEWS.md
@@ -1,3 +1,12 @@
+# ppgm 1.0.2
+
+* Fixed plot.TraitGram not plotting
+* Fixed MESS maps not including fossils
+* Modified so that input fossils can include min and max age
+* Added layerAge argument to functions to add user to input paleoclimate where layer does not equal age
+* Adds fossils to tree based on min and max age within ppgm and ppgmConsensus functions
+* Fixed examples
+
 # ppgm 1.0.1
 
 * Fixed issues for CRAN resubmission